size t

+++ src/ft_strlen.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
-size_t ft_strlen(const char *s)
+size_t  ft_strlen(const char *s)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while(s[i])
